python-generators-0x00/requirements.txt
mysql-connector-python==9.0.0
python-dotenv
